### 2025-10-23 #9235 @kthyng
@aranhax
* Can your package be made available on PyPI or conda-forge?
* I suggest adding some notebooks to demonstrate your examples, which can be added to your docs
