### 2025-11-08 a271304
Updated documentation
- updated documentation related to installation using pip

- correct doi errors on bib file

- updated documentation related to notebook with example
