+++ paper/paper.bib
@@ -80,7 +80,7 @@
     volume = {99},
     number = {D7},
     pages = {14415-14428},
-    doi = {https://doi.org/10.1029/94JD00483},
+    doi = {10.1029/94JD00483},
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/94JD00483},
     eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/94JD00483},
     year = {1994}
@@ -118,7 +118,7 @@
     pages = {104728},
     year = {2020},
     issn = {1364-8152},
-    doi = {https://doi.org/10.1016/j.envsoft.2020.104728},
+    doi = {10.1016/j.envsoft.2020.104728},
     url = {https://www.sciencedirect.com/science/article/pii/S136481521931165X},
     author = {James R. Craig and Genevieve Brown and Robert Chlumsky and R. Wayne Jenkinson and Georg Jost and Konhee Lee and Juliane Mai and Martin Serrer and Nicholas Sgro and Mahyar Shafii and Andrew P. Snowdon and Bryan A. Tolson},
 }
@@ -165,7 +165,7 @@
     year = {2013},
     note = {Snow–Atmosphere Interactions and Hydrological Consequences},
     issn = {0309-1708},
-    doi = {https://doi.org/10.1016/j.advwatres.2012.07.013},
+    doi = {10.1016/j.advwatres.2012.07.013},
     url = {https://www.sciencedirect.com/science/article/pii/S0309170812002011},
     author = {Richard Essery and Samuel Morin and Yves Lejeune and Cécile {B Ménard}},
 }
