### 2025-11-08 #9235 @aranhax
@kthyng

* I added a notebook with an example.
* pysnowclim is now available on PyPI.
* The suggestion of reviewers based on the database are: kellykochanski, Simon Filhol, Irene Garousi-Nejad, tgoelles and Martin Vonk
