### 2026-01-19 9320cb7
Modified paper to accommodate JOSS changes
Sections added:
- Software Design

- Research Impact Statement

- AI usage disclosure
