### 2026-01-20 604ae17
Small update on text
Small update on the AI section.
