+++ paper/paper.md
@@ -21,7 +21,7 @@ affiliations:
    index: 1
  - name: Woodwell Climate Research Center, Falmouth, MA 02540, USA
    index: 2
-date: 19 January 2026
+date: 20 January 2026
 bibliography: paper.bib
 ---
 
@@ -98,9 +98,7 @@ and further development.
 An example is the current implementation of coupling between
 `pySnowClim` and the Community Water Model (CWatM) [@BurekCWatM2020]
 [(https://github.com/iiasa/CWatM)](https://github.com/iiasa/CWatM).
-In addition, a new functionality was added to reduce excessive snow accumulation
-(i.e. snow towers) using an optional radiation enhancement makgin the model more realistic.
-
+In addition, to make the model more realistic, a new functionality was added to reduce excessive snow accumulation (i.e. snow towers) using an optional radiation enhancement.
 
 
 # The model
@@ -161,9 +159,9 @@ melting conditions, and seasonal variations with options for different complexit
 
 # AI usage disclosure
 
-During the preparation of this work the authors used OpenAI’s ChatGPT, and Antrhopic's Claude in order to receive recommended code feedback and edits.
-In addition the documentation of the package was originally generated by Claude.
-After using these tools, the authors reviewed and edited the content as needed and take full responsibility for the content of this publication. 
+In preparing this work, the authors used ChatGPT (OpenAI) and Claude (Anthropic) to assist with the 
+initial MATLAB‑to‑Python translation and generation of code documentation.  The unit tests of the package were carried out by using Claude.
+After using these tools, the authors thoroughly reviewed, edited and refined the content, and take full responsibility for the content of this publication.
 No generative AI tools were used in the development of the writing of this manuscript.
 
 # Acknowledgements
