### 2026-01-26 #9235 @jatkinson1000
Hi @aranhax I will be editing this paper with you.

Thank you for your patience over the holiday period and whilst we worked down the backlog.

I just did a quick check and your paper still needs a "State of the Field" section.
This isn't an oversight on your part, it seems that this check was missing from the new editorialbot and automated GitHub comment on our end. You can see the guidance on the website here: [What should my paper contain?](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain).

If you could add this (from reading the paper you likely have the information in there already and it just needs a restructure) and let me know we can proceed towards review.

In the meantime I will have a closer read of the paper, repository, and start lining up potential reviewers.

Again, thank you for your patience as JOSS goes through this change and everything settles.
