### 2026-02-04 e3fbc32
Update section paper
Paper was reorganized to include the section State of the field.
