### 2026-02-06 #9235 @jatkinson1000
Thanks @aranhax this looks good to me now.
I'll proceed next with inviting reviewers.
