### 2026-02-06 #9235 @jatkinson1000
Hi there @igarousi, @tgoelles, @JordiBolibar, and @facusapienza21!

[This paper](https://github.com/openjournals/joss-reviews/issues/9235#issuecomment-3858509551) has been submitted to JOSS, the Journal of Open Source Software.
It is [a physical snowpack model](https://github.com/abbylute/pySnowClim) written in Python.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) (note that these were recently updated with [JOSS scope changes](https://github.com/openjournals/joss-reviews/issues/9235#issuecomment-3746120908)).
