### 2026-03-13 #9235 @jatkinson1000
Thanks @gampnico - I'll give it a couple of days next week to see if we get a third reviewer (JOSS preference) but if not I'll start the review. Feel free to start checking to software and reading the paper on Monday however.
