### 2026-03-16 #10191 @editorialbot
[REVIEW]: pySnowClim: fast process-based snow modeling in Python
**Submitting author:** @aranhax (<a href="http://orcid.org/0000-0003-2514-891X">Aranildo Lima</a>)
**Repository:** https://github.com/abbylute/pySnowClim
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
