### 2026-03-27 #10191 @jatkinson1000
Hi @niccolozanotti, @gampnico, @sjavis,

I can see we have checklists generated, but it doesn't seem like much has been done beyond the basic checks so far.
Please could you all attempt to work through these over the next 2 weeks and raise any areas of improvement for the authors.
JOSS likes to aim for a 4 week turnaround where possible, including time for authors to address feedback and make any required improvements.

If you will need longer than this for any reason please let me know so that I, the author, and the other reviewers are aware.
