### 2026-04-10 #10191 @jatkinson1000
Hi @niccolozanotti and @gampnico,
You were both aiming to work through this by the end of this week.
Are we still on track for that, and could you perhaps make this a priority so that we can help move things forward and the author can start addressing any issues/comments you might raise?

I know @sjavis is currently unavailable, but they can hopefully update us next week.
