### 2026-04-10 abbylute/pySnowClim#80 @gampnico
Review comments for JOSS submission
Dear all,

These are my review comments as part of the submission process for JOSS, for the [paper here](https://github.com/abbylute/pySnowClim/blob/6a641c479d8607836054f52976bbb0f4c9914d0d/paper/paper.md).

Overall, the quality of the software and its documentation is excellent. However, I found some of the sections in the paper do not pass JOSS criteria. I am currently grading this as **3: major revisions needed**. This is not a rejection, but a request for futher input from the authors.

# General checks

Most of the criteria here are met. `pySnowClim` does not appear to have a tagged release, and no version number is given in the paper. Does the paper refer to v0.1.0 on PyPI, or the latest main branch on Github?

## Scope and significance

I have flagged this to the editor. The criteria for scope and significance are:

- Published research citing or using the software: no.
- Evidence of adoption by other research groups: tentative yes, (CWaTM). This is listed in CWaTM's documentation, but this is not linked in the paper.
- Integration into established research workflows or pipelines: tentative yes, evidence of one PR for CWaTM open since last year.
- Demonstrated performance improvements or novel capabilities compared to alternatives: no. No benchmarks provided. No explanation of how the optional radiation enhancement works, or figures comparing it to SnowClim 1.0.
- Clear potential for reuse and citation by other researchers in the domain: yes.

This means I personally do not consider the criteria are met within the paper's text, and the paper could benefit from additional text explaining how pySnowClim meets these criteria.

# Development history and open-source practice

**Collaborative effort: Good**

## Development timeline (OK)

Whilst the repository was created in 2024, it appears ~60% of commits occur in over a single month last year. However, the code for the original SnowClim model was published in 2021 by one of the authors. I would consider this requirement satisfied if taken in context with the wider development of the SnowClim ecosystem.

## Open development (OK)

No evidence of releases or external engagement, but a long history of pull requests.

## Good practices  (Good)

Code coverage is at 84%. Includes CI. However, there appear to be no tagged versions or releases, and the version number is still v0.1.0, which semantically suggests this software is in an alpha state, rather than a finished release.

# Functionality (Good)

All criteria met. Installation was very simple.

# Documentation (Good)

All criteria met. I was very impressed by the clarity of the documentation. There were some minor discrepancies between the documented output variable names and the actual output files, but these did not affect my understanding of the software.

# Software paper content.

## Statement of need (OK)

An additional paragraph explaining how pySnowClim relates to other work in the field.

## State of the field (OK)

The authors mention "current snow models", but the most recent paper they cite dates from 2021, and most of their citations for this section are from the mid-2000s. Can the authors provide more current examples of snow models?

What motivated the authors to rewrite SnowClim in Python, i.e. what research gap does pySnowClim fill that is not already addressed by SnowClim?

## Software design (Bad)

Design is described as following "object-oriented design principles", but encapsulation is limited to snowpack/snow model variables and precipitation properties. The majority of the code is MATLAB-style functional programming with separate modules for a single function. Perhaps the authors could rewrite this section to outline why they chose this particular design approach?

Using numpy/pandas/xarray in Python models is fairly standard by now, and overall the description of the software's design seems quite generic. Phrases like "enhancing computational efficiency" and "more efficient" could benefit from benchmarks against the MATLAB version of SnowClim, or at the very least a comparison against other packages. Did the authors encounter any issues or workarounds when using these packages, and if so, what design choices did they make to resolve these?

I'm unsure why the authors include a description of the documentation, as this is not a software design.

To better meet JOSS' criteria, I would suggest adding a section on any trade-offs made during design, and why the authors' design choices are relevant for their research application.

## Research impact statement (Bad)

The research impact statement is brief and appears mostly aspirational, focusing on the implied benefits of Python over MATLAB rather than the model itself.

Following [JOSS guidelines here](https://joss.readthedocs.io/en/latest/review_criteria.html#research-impact-statement-required), the authors do not appear to provide compelling evidence for realised impact in terms of publications and evidence of external use.

Authors do mention that pySnowClim has been coupled to CWatM as evidence of integration with other research tools, but the [pull request is still open since last year](https://github.com/iiasa/CWatM/pull/203). The CWaTM documentation mentions this coupling is still being tested. I think here a reader would benefit from a clear and direct link, perhaps the relevant documentation page for CWaTM.

Is there a recent, concrete example of integration with other tools, i.e. a study, blog post, student thesis, Jupyter notebook?

The paper also does not provide evidence of credible near-term significance. The authors mention the model is "more realistic", presumably referring to SnowClim 1.0 (and this is not directly stated). Can the authors provide any benchmarks or data to support this statement? Are there examples of reproducible research materials demonstrating how pySnowClim is used, beyond documentation examples with test data?

> a new functionality was added to reduce excessive snow accumulation (i.e. snow towers) using an optional radiation enhancement.

If this is a novel feature, are the authors able to show this is more accurate than existing methods (benchmarks, statistics, figures)? If this is not novel, please include a reference.

## Quality of writing (Good)

The paper is generally well written. I would perhaps suggest adding a clear explanation of how the authors implemented snow towers through the optional radiation enhancement, since the paper mentions this is a new feature.

## References

> a new functionality was added to reduce excessive snow accumulation (i.e. snow towers) using an optional radiation enhancement.

If this is not a novel feature or method, please provide a reference.
