### 2026-04-10 #10191 @gampnico
Hi @jatkinson1000. I've finished my review, and when following JOSS' guidelines I would grade this submission as **3: major revisions needed**. I have submitted [comprehensive review comments](https://github.com/abbylute/pySnowClim/issues/80) via their repository's issue tracker. JOSS guidelines weren't very clear if this is the expected way of doing things.

I would however flag the scope and significance of the paper, as this is a Python port of the existing SnowClim model by one of the authors, with one (documented) additional improvement. Following the review criteria:

- Published research citing or using the software: no.
- Evidence of adoption by other research groups: tentative yes, (CWaTM). This is listed in CWaTM's documentation, but this is not linked in the paper.
- Integration into established research workflows or pipelines: tentative yes, evidence of one PR for CWaTM still open since last year.
- Demonstrated performance improvements or novel capabilities compared to alternatives: no.
- Clear potential for reuse and citation by other researchers in the domain: yes.

Please also see my review comment on their research impact statement, but I'd prefer to defer to you on whether this meets JOSS' criteria for significance, as it may be that the authors just need to add more text.
