### 2026-04-13 #10191 @niccolozanotti
@jatkinson1000
Apologies to the authors and the editor for slightly going past the announced review time window.

I've just completed my review of the paper submitted and I find myself in agreement with the other reviewer in marking this as needing **major revisions**, especially regarding the software paper.

The following are my comments:

**Package assessment**

* The package installation was straightfoward and the basic claimed functionality was easy to  check via automated tests and examples.
* The repository shows sustained development and open development practices. The git history includes a substantial sequence of pull-request merges and multiple contributor identities, which is enough for me to consider the open-development and collaborative-effort criteria satisfied, even if external community involvement still appears limited.
* I did not find any tagged git releases, which would be worth addressing for archival clarity, but this does not by itself block JOSS review.

**Software Paper**
* The strongest issue is that several required JOSS paper sections are present only in a minimal form.
* The **State of the field** section is too general. It describes a gap between simple and process-based snow models, but it does not compare `pySnowClim` concretely against currently used alternatives, nor does it make a strong "build vs. contribute" case.
Since this package is a Python implementation of an existing SnowClim model, the paper should be much more explicit about what is genuinely new here beyond language porting, packaging, and one added option for reducing excessive snow accumulation.
* The **Software Design** section does not yet explain design trade-offs in a compelling way. At present it mainly lists technologies (`NumPy`, `xarray`, `pandas`) and broad architectural claims. For JOSS, I would expect a clearer explanation of what design decisions were made, what alternatives were considered or rejected, and why the chosen design matters scientifically or operationally. For example, if vectorized gridded execution, modular physics functions, or API choices materially improve extensibility, coupling, or reproducibility, that should be argued explicitly.
* The **Research Impact Statement** is also not yet convincing. The current text is mostly prospective. The paper mentions coupling with CWatM, but as correctly pointed out by the other reviewer this is still a work in progress. If there are no published downstream studies yet, then the paper should instead provide stronger evidence such as reproducible benchmarks, adoption in teaching or workflows, comparisons to the MATLAB implementation, or clearer evidence that the new Python package materially expands the model's usability and reuse.

* The paper is understandable, but it still needs editing for wording and precision. There are several small language issues, and some claims are too broad without direct support. Examples include statements about computational efficiency, accessibility, and realism that would be stronger if either quantified or narrowed.
* As for the references, the state-of-the-field discussion relies on a relatively dated set of examples, and the statement about the optional radiation enhancement used to reduce "snow towers" appears to need either a supporting citation or a clearer explanation that this is a new contribution of the package.
