### 2026-04-17 #10191 @sjavis
@jatkinson1000 I have finished my review. Apologies for the delay in completing this.

Overall, I would grade this paper as needing **minor revisions**. The state of the software and documentation is good, but there are some issues in the paper for which I have suggested improvements below. The main blocking point is the software design section in my opinion.

**Statement of need** - OK. This could have more about the gap it fills within the field. Although this is somewhat covered in the following section.
**State of the field** - OK. It would ideally be a bit more specific in the comparison with existing tools.
**Software design** - OK/Bad. While there is some rationale for the software architectural choices it is not very detailed and should include more reasoning for why the choices matter.
**Research impact statement** - OK. The integration with the Community Water Model is good, but it would ideally have additional citations or evidence for external use, even if this was only for the original MATLAB model.
**AI usage disclosure** - Good.
**References** - The citations and format seem good. As stated in my comment about the research impact statement, it would ideally have additional citations for the use of the software in research.
