+++ paper/paper.bib
@@ -178,3 +178,93 @@
     url={https://hydrology.usu.edu/dtarb/snow/snowrep.pdf},
     note={Last accessed: 2025-09-22}
 }
+
+ @article{Anderson2023plant, 
+    title={The consequences of winter climate change for plant performance}, 
+    volume={110}, 
+    ISSN={0002-9122, 1537-2197}, 
+    DOI={10.1002/ajb2.16252}, 
+    number={12}, 
+    journal={American Journal of Botany}, 
+    author={Anderson, Jill T.}, 
+    year={2023}, 
+    month=dec, 
+    pages={e16252}, 
+    language={en}
+}
+
+  @article{Jans2025Sentinel, 
+    title={Sensitivity of Sentinel-1 C-band SAR backscatter, polarimetry and interferometry to snow accumulation in the Alps}, 
+    volume={316}, 
+    ISSN={0034-4257}, 
+    DOI={10.1016/j.rse.2024.114477}, 
+    journal={Remote Sensing of Environment}, 
+    author={Jans, Jonas-Frederik and Beernaert, Ezra and De Breuck, Morgane and Brangers, Isis and Dunmire, Devon and De Lannoy, Gabrielle and Lievens, Hans}, 
+    year={2025}, 
+    month=jan, 
+    pages={114477}
+}
+
+ @article{Anderson2025geneflow, 
+    title={Adaptation and gene flow are insufficient to rescue a montane plant under climate change}, 
+    volume={388}, 
+    DOI={10.1126/science.adr1010}, 
+    number={6746}, 
+    journal={Science}, 
+    publisher={American Association for the Advancement of Science}, 
+    author={Anderson, Jill T. and DeMarche, Megan L. and Denney, Derek A. and Breckheimer, Ian and Santangelo, James and Wadgymar, Susana M.}, 
+    year={2025}, 
+    month=may, 
+    pages={525–531}
+}
+
+ @article{Dixon2026rainonsnow, 
+    title={Numerical representation of turbulent fluxes during the March 2019 Nebraska rain-on-snow event}, 
+    volume={21}, 
+    ISSN={1748-9326}, 
+    DOI={10.1088/1748-9326/ae3fb0}, 
+    number={4}, 
+    journal={Environmental Research Letters}, 
+    publisher={IOP Publishing}, 
+    author={Dixon, Ross D and Janzon, Erik J and Roy, Tirthankar and Suriano, Zachary J and Davidson, Samuel}, 
+    year={2026}, 
+    month=feb, 
+    pages={044023}, 
+    language={en}
+}
+
+ @article{Williams2026forestfire, 
+    title={The Western United States Large Forest-Fire Stochastic Simulator (WULFFSS) 1.0: a monthly gridded forest-fire model using interpretable statistics}, 
+    volume={19}, 
+    ISSN={1991-959X}, 
+    DOI={10.5194/gmd-19-1157-2026}, 
+    number={3}, 
+    journal={Geoscientific Model Development}, 
+    publisher={Copernicus GmbH}, 
+    author={Williams, A. Park and Hansen, Winslow D. and Juang, Caroline S. and Abatzoglou, John T. and Radeloff, Volker C. and Wang, Bowen and Hall, Jazlynn and Buch, Jatan and Madakumbura, Gavin D.}, 
+    year={2026}, 
+    month=feb, 
+    pages={1157–1191}, 
+    language={English}
+}
+
+@article{Boeykens2025MLsentinel, 
+    title={Machine learning for snow depth estimation over the European Alps, using Sentinel-1 observations, meteorological forcing data and physically-based model simulations}, 
[…]
