### 2026-04-23 abbylute/pySnowClim#80 @aranhax
Hi @gampnico thank you for the suggestions.

I updated the paper following some of your suggestions. Unfortunately because of the size limitation of the paper we couldn't address all the suggestions/comments there. Therefore, I will use this issue to answer the remaining ones.

1- Most of the criteria here are met. pySnowClim does not appear to have a tagged release, and no version number is given in the paper. Does the paper refer to v0.1.0 on PyPI, or the latest main branch on Github?

They are the same version PyPI, and the latest main.  The latest changes I had to perform on the main was to release on PyPI.

2- Published research citing or using the software: no.

Additional references of the original MATLAB version of snowclim were added on the Research impact statement section .

3- Evidence of adoption by other research groups: tentative yes, (CWaTM). This is listed in CWaTM's documentation, but this is not linked in the paper.

pySnowclim code was already incorporated on the CWatM develop branch (https://github.com/iiasa/CWatM/tree/develop/cwatm/hydrological_modules/pySnowClim). At this moment the CWatM developers are testing and fixing minor issues related to the coupling (https://github.com/iiasa/CWatM/tree/develop).

4- Integration into established research workflows or pipelines: tentative yes, evidence of one PR for CWaTM open since last year.

CWatM is a relatively new hydrological model with active development community and a very modular code. However, for coupling pySnowClim with CWatMs seamlessly , several components inside the CWatM snow_frost.py module had to be modified to make it more modular. Besides snow, other hydrological routines inside CWatM have been updated in parallel, while our pull request was behind the main repo. For these reasons, the CWatM developers didn’t merge the PR but copied the code and implemented the coupling inside the snow_frost module. However, the core pySnowClim code has been integrated into CWatM with only minor changes. Going forward, there are plans to incorporate pySnowClim as a submodule directly to CWatM instead of copying the code to avoid desynchronizations between the two repos.

5- Demonstrated performance improvements or novel capabilities compared to alternatives: no. No benchmarks provided. No explanation of how the optional radiation enhancement works, or figures comparing it to SnowClim 1.0.

This is also motivated by JOSS scope “Design thinking: Meaningful architectural decisions and trade-offs considered. We particularly value work that builds upon or extends existing software ecosystems rather than reinventing solutions where quality alternatives already exist.” We have revised the State of the field, Software design and Research impact statement sections to explain the intention of pySnowClim to extend existing software ecosystem leveraging the advantages of using Python as a programing language instead of the original MATLAB version, which restricts wider applications. Thus, tests were carried out to compare if the outputs from MATLAB and Python versions were physically plausible (e.g. issue #46  - closed). The optional radiation enhancement is a simple factor also used by other models which is lacking on the original MATLAB version. At high altitudes because of their climatic conditions (e.g. temperature never exceeding 1 ◦C), it is common for some models to accumulate snow without any melting loss. In these altitudes snow is accumulated and is converted into firn, which is then converted into ice as glaciers move to lower regions over decades or even centuries. In the ablation area, the ice is again melted. In pySnowClim this process can be optionally simulated by melting the snow using the radiation enhancement on an annual basis. Some references were added to The model section. In addition figures that were generated during the tests of the functionality were added to its respective issue (#53 ).

Regarding the novel capability pySnowClim, the restructuring and translation of the SnowClim MATLAB code to Python makes it ready to be coupled with the CWatM model. The coupled pySnowClim-CWatM model enhances the capability of CWatM to provide physically based snow simulation. We have added these clarifications in the revised version of the manuscript.
