### 2026-04-23 #10191 @aranhax
Thank you @niccolozanotti and @sjavis for the suggestions.

I uploaded a new version of the paper on the repo. In addition you might find some extra clarification on the issue https://github.com/abbylute/pySnowClim/issues/80 created by @gampnico .
