### 2026-08-14 08ce40f
Update paper.md
- Added links to the research impact statement
- Modified the state of the field section
