### 2026-08-14 340045e
Update paper.bib
- Updated references of the modified state of the field section
