+++ paper/paper.bib
@@ -1,3 +1,109 @@
+@article {Wurzer2016,
+      author = "Sebastian  Würzer and Tobias  Jonas and Nander  Wever and Michael  Lehning",
+      title = "Influence of Initial Snowpack Properties on Runoff Formation during Rain-on-Snow Events",
+      journal = "Journal of Hydrometeorology",
+      year = "2016",
+      publisher = "American Meteorological Society",
+      address = "Boston MA, USA",
+      volume = "17",
+      number = "6",
+      doi = "10.1175/JHM-D-15-0181.1",
+      pages=      "1801 - 1815",
+      url = "https://journals.ametsoc.org/view/journals/hydr/17/6/jhm-d-15-0181_1.xml"
+}
+
+@article{Ismail2023,
+    AUTHOR = {Ismail, M. F. and Bogacki, W. and Disse, M. and Sch\"afer, M. and Kirschbauer, L.},
+    TITLE = {Estimating degree-day factors of snow based on energy flux components},
+    JOURNAL = {The Cryosphere},
+    VOLUME = {17},
+    YEAR = {2023},
+    NUMBER = {1},
+    PAGES = {211--231},
+    URL = {https://tc.copernicus.org/articles/17/211/2023/},
+    DOI = {10.5194/tc-17-211-2023}
+}
+
+@article{Essery2015,
+    author = {Essery, R.},
+    title = {A factorial snowpack model (FSM 1.0)},
+    journal = {Geoscientific Model Development},
+    volume = {8},
+    pages = {3867-3876},
+    year = {2015},
+    doi = {10.5194/gmd-8-3867-2015},
+    url = {https://doi.org/10.5194/gmd-8-3867-2015}
+}
+
+@article{Marks1999,
+    author = {Marks, D. and Domingo, J. and Susong, D. and Link, T. and Garen, D.},
+    title = {A spatially distributed energy balance snowmelt model for application in mountain basins},
+    journal = {Hydrological Processes},
+    volume = {13},
+    pages = {1935-1959},
+    year = {1999},
+    doi = {10.1002/(SICI)1099-1085(199909)13:12/13<1935::AID-HYP868>3.0.CO;2-C},
+    url = {https://doi.org/10.1002/(SICI)1099-1085(199909)13:12/13<1935::AID-HYP868>3.0.CO;2-C}
+}
+
+@techreport{Jordan1991,
+    author = {Jordan, Rachel E.},
+    title = {A One-dimensional temperature model for a snow cover : technical documentation for SNTHERM.89},
+    series = {CRREL Special Report},
+    number = {91-16},
+    year = {1991},
+    url = {https://erdc-library.erdc.dren.mil/server/api/core/bitstreams/81b728f8-8f72-4ef8-e053-411ac80adeb3/content}
+}
+
+@article{Lehning1999,
+    author = {Lehning, Michael and Bartelt, Perry and Brown, Bob and Russi, Tom and Stockli, Urs and Simmerli, Martin},
+    title = {snowpack model calculations for avalanche warning based upon a new network of weather and snow stations},
+    journal = {Cold Regions Science and Technology},
+    volume = {30},
+    pages = {145-157},
+    year = {1999},
+    doi = {10.1016/S0165-232X(99)00022-1},
+    url = {https://doi.org/10.1016/S0165-232X(99)00022-1}
+}
+
+@techreport{Bergstrom1992,
+    author = {Bergstrom, Sten},
+    title = {The HBV Model- its structure and applications},
+    series = {SMHI Reports Hydrology},
+    number = {4},
+    year = {1992},
+    url = {https://www.smhi.se/en/publications-from-smhi/publications/2015-01-13-the-hbv-model---its-structure-and-applications.}
+}
[…]
