### 2026-08-14 c9dc2bd
Update paper.md
Small fix to generate the pdf and [tag review](https://github.com/openjournals/joss-reviews/issues/10191#issuecomment-5057329164)
