### 2026-08-14 2626fe1
Update paper.bib
Fix pdf build.
