+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article {Wurzer2016,
-      author = "Sebastian  Würzer and Tobias  Jonas and Nander  Wever and Michael  Lehning",
+      author = "Sebastian  W\"rzer and Tobias  Jonas and Nander  Wever and Michael  Lehning",
       title = "Influence of Initial Snowpack Properties on Runoff Formation during Rain-on-Snow Events",
       journal = "Journal of Hydrometeorology",
       year = "2016",
