### 2026-08-14 ef9b7ea
Update paper.bib
Fix pdf build
