+++ paper/paper.bib
@@ -1,29 +1,3 @@
-@article {Wurzer2016,
-      author = "Sebastian  W\"rzer and Tobias  Jonas and Nander  Wever and Michael  Lehning",
-      title = "Influence of Initial Snowpack Properties on Runoff Formation during Rain-on-Snow Events",
-      journal = "Journal of Hydrometeorology",
-      year = "2016",
-      publisher = "American Meteorological Society",
-      address = "Boston MA, USA",
-      volume = "17",
-      number = "6",
-      doi = "10.1175/JHM-D-15-0181.1",
-      pages=      "1801 - 1815",
-      url = "https://journals.ametsoc.org/view/journals/hydr/17/6/jhm-d-15-0181_1.xml"
-}
-
-@article{Ismail2023,
-    AUTHOR = {Ismail, M. F. and Bogacki, W. and Disse, M. and Sch\"afer, M. and Kirschbauer, L.},
-    TITLE = {Estimating degree-day factors of snow based on energy flux components},
-    JOURNAL = {The Cryosphere},
-    VOLUME = {17},
-    YEAR = {2023},
-    NUMBER = {1},
-    PAGES = {211--231},
-    URL = {https://tc.copernicus.org/articles/17/211/2023/},
-    DOI = {10.5194/tc-17-211-2023}
-}
-
 @article{Essery2015,
     author = {Essery, R.},
     title = {A factorial snowpack model (FSM 1.0)},
