### 2026-08-14 76e850a
Update paper.bib
Fix pdf build
