### 2026-08-14 ac4400c
Update paper.bib
New pdf
