+++ paper/paper.bib
@@ -1,3 +1,61 @@
+@techreport{Jordan1991,
+    author = {Jordan, Rachel E.},
+    title = {A One-dimensional temperature model for a snow cover : technical documentation for SNTHERM.89},
+    series = {CRREL Special Report},
+    number = {91-16},
+    year = {1991},
+    url = {https://erdc-library.erdc.dren.mil/server/api/core/bitstreams/81b728f8-8f72-4ef8-e053-411ac80adeb3/content}
+}
+
+@article{Lehning1999,
+    author = {Lehning, Michael and Bartelt, Perry and Brown, Bob and Russi, Tom and Stockli, Urs and Simmerli, Martin},
+    title = {snowpack model calculations for avalanche warning based upon a new network of weather and snow stations},
+    journal = {Cold Regions Science and Technology},
+    volume = {30},
+    pages = {145-157},
+    year = {1999},
+    doi = {10.1016/S0165-232X(99)00022-1},
+    url = {https://doi.org/10.1016/S0165-232X(99)00022-1}
+}
+
+@techreport{Bergstrom1992,
+    author = {Bergstrom, Sten},
+    title = {The HBV Model- its structure and applications},
+    series = {SMHI Reports Hydrology},
+    number = {4},
+    year = {1992},
+    url = {https://www.smhi.se/en/publications-from-smhi/publications/2015-01-13-the-hbv-model---its-structure-and-applications.}
+}
+
+@article{Arnold1998,
+    author = {Arnold, J.G. and Srinivasan, R., and Muttiah, R.S. and Williams, J.R.},
+    title = {Large Area Hydrologic Modeling and Assessment Part 1: Model Development},
+    journal = {Journal of the American Water Resources Association},
+    volume = {34},
+    pages = {73-89},
+    year = {1998},
+    doi = {10.1111/j.1752-1688.1998.tb05961.x},
+    url = {https://doi.org/10.1111/j.1752-1688.1998.tb05961.x}
+}
+
+@book{Martinec2008,
+    author = {Martinec, J. and Rango, A. and Roberts, R.},
+    title = {Snowmelt Runoff Model (SRM) User's Manual},
+    year = {2008},
+    publisher location = {Las Cruces, New Mexico},
+    publisher = {New Mexico State University},
+    url = {https://www.ars.usda.gov/research/publications/publication/?seqNo115=228834}
+}
+
+@techreport{Anderson1973,
+    author = {Anderson, Eric A.},
+    title = {National Weather Service river forecast system: snow accumulation and ablation model},
+    series = {NOAA technical memorandum NWS HYDRO},
+    number = {17},
+    year = {1973},
+    url = {https://repository.library.noaa.gov/view/noaa/13507}
+}
+
 @incollection{Caretta2022Water,
     author = {Caretta, M.A. and Mukherji, A. and Arfanuzzaman, M. and Betts, R.A. and Gelfan, A. and Hirabayashi, Y. and Lissner, T.K. and Liu, J. and Lopez Gunn, E. and Morgan, R. and Mwanga, S. and Supratid, S.},
     title = {Water},
