### 2026-08-14 e2ffa37
Update paper.bib
Fix pdf
