+++ paper/paper.bib
@@ -1,23 +1,3 @@
-@techreport{Jordan1991,
-    author = {Jordan, Rachel E.},
-    title = {A One-dimensional temperature model for a snow cover : technical documentation for SNTHERM.89},
-    series = {CRREL Special Report},
-    number = {91-16},
-    year = {1991},
-    url = {https://erdc-library.erdc.dren.mil/server/api/core/bitstreams/81b728f8-8f72-4ef8-e053-411ac80adeb3/content}
-}
-
-@article{Lehning1999,
-    author = {Lehning, Michael and Bartelt, Perry and Brown, Bob and Russi, Tom and Stockli, Urs and Simmerli, Martin},
-    title = {snowpack model calculations for avalanche warning based upon a new network of weather and snow stations},
-    journal = {Cold Regions Science and Technology},
-    volume = {30},
-    pages = {145-157},
-    year = {1999},
-    doi = {10.1016/S0165-232X(99)00022-1},
-    url = {https://doi.org/10.1016/S0165-232X(99)00022-1}
-}
-
 @techreport{Bergstrom1992,
     author = {Bergstrom, Sten},
     title = {The HBV Model- its structure and applications},
