### 2026-08-14 45d8726
Update paper.bib
Fix pdf
