+++ paper/paper.bib
@@ -1,23 +1,3 @@
-@techreport{Bergstrom1992,
-    author = {Bergstrom, Sten},
-    title = {The HBV Model- its structure and applications},
-    series = {SMHI Reports Hydrology},
-    number = {4},
-    year = {1992},
-    url = {https://www.smhi.se/en/publications-from-smhi/publications/2015-01-13-the-hbv-model---its-structure-and-applications.}
-}
-
-@article{Arnold1998,
-    author = {Arnold, J.G. and Srinivasan, R., and Muttiah, R.S. and Williams, J.R.},
-    title = {Large Area Hydrologic Modeling and Assessment Part 1: Model Development},
-    journal = {Journal of the American Water Resources Association},
-    volume = {34},
-    pages = {73-89},
-    year = {1998},
-    doi = {10.1111/j.1752-1688.1998.tb05961.x},
-    url = {https://doi.org/10.1111/j.1752-1688.1998.tb05961.x}
-}
-
 @book{Martinec2008,
     author = {Martinec, J. and Rango, A. and Roberts, R.},
     title = {Snowmelt Runoff Model (SRM) User's Manual},
