### 2026-08-14 9d0ad25
Update paper.bib
Add 1 reference
