+++ paper/paper.bib
@@ -1,12 +1,3 @@
-@book{Martinec2008,
-    author = {Martinec, J. and Rango, A. and Roberts, R.},
-    title = {Snowmelt Runoff Model (SRM) User's Manual},
-    year = {2008},
-    publisher location = {Las Cruces, New Mexico},
-    publisher = {New Mexico State University},
-    url = {https://www.ars.usda.gov/research/publications/publication/?seqNo115=228834}
-}
-
 @techreport{Anderson1973,
     author = {Anderson, Eric A.},
     title = {National Weather Service river forecast system: snow accumulation and ablation model},
