### 2026-08-14 92dae04
Update paper.bib
Test pdf
