+++ paper/paper.bib
@@ -4,7 +4,7 @@
     year = {2008},
     publisher location = {Las Cruces, New Mexico},
     publisher = {New Mexico State University},
-    url = "{https://www.ars.usda.gov/research/publications/publication/?seqNo115=228834}"
+    url = {"https://www.ars.usda.gov/research/publications/publication/?seqNo115=228834"}
 }
 
 @techreport{Anderson1973,
