### 2026-08-14 f31dc02
Update paper.bib
Fixed bib
