### 2026-08-14 e5d1920
Update paper.md
Fixed reference.
