### 2026-08-19 d364646
Fix turbulent-flux bugs (issue #84)
- Fix E0_app==2 branch: compare parameters['E0_app'] (was comparing the
  literal list ['E0_app'] to 2, so the branch was always False/dead code).
- Compute the bulk Richardson number (Rib) unconditionally instead of only
  when parameters['stability'] is on. Rib is also needed by the
  E0_stable==2 windless-exchange gating, which previously raised
  NameError whenever stability=0 was combined with E0_stable=2 since Rib
  was never computed in that case. The stability-only FH adjustment to CH
  remains gated on parameters['stability'] as before; this only ch
[…]
