### 2026-08-24 f6fc979
removed changes to snowclim_model.py
Removed the changes to snowclim_model.py that are already included in PR #83
